requirements2.txt
numpy
pandas
tensorflow
tensorflow-gpu
sklearn
keras
Pillow
tqdm
matplotlib